Only push to OSSRH for promote builds
Previously the publish script used the presence of
SNAPSHOT in the VERSION file to determine whether
jars should be pushed to OSSRH. This was unreliable
as SNAPSHOT is stripped from the VERSION file before
the release block is called.

This commit uses the MODE environment variable instead.
Jars are only pushed if MODE=PROMOTE.