Add Release to central from OSSRH
Pushing to sonatype oss is not enough for a package to
be included in maven central. To do this nexus-staging:release
must be called via maven.

This commit also fixes an error that caused the release process
to fail as it attempted to create a tag that contained a newline.